fix when revise=false, small refactor
