+++ config.cfg
@@ -1,6 +1,7 @@
 [OpenAI]
 # Optional. Set these if you are testing OpenAI models.
 api_key = 
+
 # Optional. Set this if you are using an alternative OpenAI compatible endpoint, like ollama running locally
 # openai_compatible_url = http://localhost:11434/v1/
 
@@ -20,8 +21,6 @@ google_spreadsheet_url =
 [Options]
 # Set to true or false
 trust_remote_code = false
-# You can specify an alternative language for question dataset and scoring here. Currently supported: de = German
-language = de
 
 [Oobabooga config]
 # e.g. ~/text-generation-webui/start_linux.sh
